Removed wav buffer
I removed the need to buffer the audio in a wav file by changing some formats and deploying code in numpy to map the buffer directly from pyaudio to pyloudnorm. However, there seems to be a difference with main and this branch, and it has to do with being in callback mode, as everything else is basically identical. Can't accept the PR until the meter values are accurate to WLM standard